% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdt.R
\name{cdt_pslg}
\alias{cdt_pslg}
\title{Convert an RTriangle pslg to cdt_triangulate arguments}
\usage{
cdt_pslg(p, ...)
}
\arguments{
\item{p}{a pslg}

\item{...}{passed to cdt_triangulate}
}
\description{
Convert an RTriangle pslg to cdt_triangulate arguments
}
