% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdt.R
\name{cdt_pslg}
\alias{cdt_pslg}
\title{Triangulate an RTriangle pslg with the CDT backend}
\usage{
cdt_pslg(p, ...)
}
\arguments{
\item{p}{a pslg (list with P and S)}

\item{...}{passed to cdt_triangulate}
}
\description{
Triangulate an RTriangle pslg with the CDT backend
}
