% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdtr-package.R
\docType{package}
\name{cdtr-package}
\alias{cdtr}
\alias{cdtr-package}
\title{cdtr: Constrained Delaunay Triangulation with Refinement via CDT}
\description{
Binding to the artem-ogre/CDT header-only C++ library (MPL-2.0): constrained and conforming Delaunay triangulation, Ruppert refinement by area or minimum angle, resolution of crossing constraints, and topological region depth per triangle. Aimed at replacing RTriangle in anglr.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hypertidy/cdtr}
  \item Report bugs at \url{https://github.com/hypertidy/cdtr/issues}
}

}
\author{
\strong{Maintainer}: Michael Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
